Support in-flight messages, align data types, removed unnecessary abstractions

+++ src/KurrentDB.SecondaryIndexing/Indexes/Category/CategorySql.cs
@@ -21,17 +21,17 @@ internal static class CategorySql {
 			from idx_all where category=$1 and category_seq>=$2 and category_seq<=$3
 			"""u8;
 
-		public static CategoryRecord Parse(ref DataChunk.Row row) => new(row.ReadInt32(), row.ReadInt64(), row.ReadInt32());
+		public static CategoryRecord Parse(ref DataChunk.Row row) => new(row.ReadInt64(), row.ReadInt64(), row.ReadInt64());
 	}
 
-	public record struct CategoryIndexQueryArgs(int Id, long FromSeq, long ToSeq);
+	public record struct CategoryIndexQueryArgs(long Id, long FromSeq, long ToSeq);
 
-	public record struct CategoryRecord(int CategorySeq, long LogPosition, int EventNumber);
+	public record struct CategoryRecord(long CategorySeq, long LogPosition, long EventNumber);
 
 	public struct GetCategoriesQuery : IQuery<ReferenceRecord> {
 		public static ReadOnlySpan<byte> CommandText => "select id, name from category"u8;
 
-		public static ReferenceRecord Parse(ref DataChunk.Row row) => new(row.ReadInt32(), row.ReadString());
+		public static ReferenceRecord Parse(ref DataChunk.Row row) => new(row.ReadInt64(), row.ReadString());
 	}
 
 	public struct GetCategoriesMaxSequencesQuery : IQuery<(long Id, long Sequence)> {
@@ -41,7 +41,7 @@ internal static class CategorySql {
 		public static (long Id, long Sequence) Parse(ref DataChunk.Row row) => (row.ReadInt64(), row.ReadInt64());
 	}
 
-	public record struct AddCategoryStatementArgs(int Id, string Category);
+	public record struct AddCategoryStatementArgs(long Id, string Category);
 
 	public struct AddCategoryStatement : IPreparedStatement<AddCategoryStatementArgs> {
 		public static BindingContext Bind(in AddCategoryStatementArgs args, PreparedStatement statement)
